Proptest key/value set index compare to rust set [ECR-2042] (#903)
